sua gio hang
them khach hang ben nhan vien va them nhan vien sua css

+++ Audio_TripleV/src/main/java/com/example/demo/controller/ArticleController.java
@@ -14,9 +14,12 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import java.util.Date;
 import java.util.List;
 
 @Controller
@@ -34,6 +37,7 @@ public class ArticleController {
 
     @Autowired
     GioHangService gioHangService;
+
 
     // Hiển thị tất cả các bài viết và bình luận tương ứng
     @GetMapping("/bai-viet/hien-thi")
@@ -59,7 +63,7 @@ public class ArticleController {
     // Hiển thị chi tiết bài viết cùng bình luận của nó
     @GetMapping("/bai-viet/{id}")
     public String showArticleDetail(@AuthenticationPrincipal UserDetails userDetails,
-                                    @PathVariable("id") Long articleId, Model model) {
+                                    @PathVariable("id") Integer articleId, Model model) {
         // Lấy thông tin khách hàng
         KhachHang khachHang = khachHangService.findByTaiKhoan(userDetails.getUsername())
                 .orElseThrow(() -> new RuntimeException("Khách hàng không tồn tại"));
